backend/requirements.txt

# Project Python dependencies (one per line). Install with:
#   pip install -r requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-decouple
djangorestframework-simplejwt
Pillow
requests
python-dotenv


